Removed `FuncOpaque::wrap*()` in favor of `FuncOpaque::new()`
Corresponding closures moved to `Function::new*()`.
